Crate fvm_sdk

Modules

actor

crypto

debug

error

event

gas

ipld

message

network

rand

send

sself

sys

This module defines the low-level syscall API.

vm

Macros

fvm_syscalls

Generate a set of FVM syscall shims.

Constants

NO_DATA_BLOCK_ID

BlockID representing nil parameters or return data.

Functions

initialize

Initialize the FVM SDK. Calling this function optional but encouraged.

Type Aliases

SyscallResult

SDK functions performing a syscall return a SyscallResult type, where the Error type is an ExitCode. ExitCode::Ok is translated to an Ok result, while error codes are propagated as Err(ExitCode).