Module fvm_sdk::actor

Functions

balance_of

Retrieves the balance of the specified actor, or None if the actor doesn’t exist.

create_actor

Creates a new actor of the specified type in the state tree, under the provided address.

get_actor_code_cid

Look up the code ID at an actor address. Returns None if the actor cannot be found.

get_builtin_actor_type

Determines whether the supplied CodeCID belongs to a built-in actor type, and to which.

get_code_cid_for_type

Returns the CodeCID for a built-in actor type. Aborts with IllegalArgument if the supplied type is invalid.

lookup_delegated_address

Looks up the delegated (f4) address of the specified actor. Returns None if the actor doesn’t exist or it doesn’t have f4 address.

next_actor_address

Generates a new actor address for an actor deployed by the calling actor.

resolve_address

Resolves the ID address of an actor. Returns None if the address cannot be resolved. Successfully resolving an address doesn’t necessarily mean the actor exists (e.g., if the addresss was already an actor ID).