Module fvm_sdk::crypto

Functions

batch_verify_seals

compute_unsealed_sector_cid

Computes an unsealed sector CID (CommD) from its constituent piece CIDs (CommPs) and sizes.

hash_blake2b

Hashes input data using blake2b with 256 bit output.

hash_into

Hashes input data using one of the supported functions into a buffer.

hash_owned

Hashes input data using one of the supported functions. hashes longer than 64 bytes will be truncated.

recover_secp_public_key

Recovers the signer public key from the message hash and signature.

verify_aggregate_seals

verify_bls_aggregate

verify_consensus_fault

Verifies that two block headers provide proof of a consensus fault:

verify_post

Verifies a window proof of spacetime.

verify_replica_update

verify_signature

Verifies that a signature is valid for an address and plaintext.