Function fvm_sdk::sys::ipld::block_link

pub unsafe fn block_link(
    id: u32,
    hash_fun: u64,
    hash_len: u32,
    cid: *mut u8,
    cid_max_len: u32
) -> Result<u32, ErrorNumber>

Computes the given block’s CID, writing the resulting CID into cid.

The returned CID is added to the reachable set.

Arguments

id is ID of the block to link.

hash_fun is the multicodec of the hash function to use.

hash_len is the desired length of the hash digest.

cid is the output buffer (in wasm memory) where the FVM will write the resulting cid.

cid_max_length is the length of the output CID buffer.

Returns

The length of the CID.

Errors

Error	Reason

InvalidHandle	if the handle isn’t known.

IllegalCid	hash code and/or hash length aren’t supported.

BufferTooSmall	if the passed buffer is too small

IllegalArgument	if the passed buffer isn’t valid, in memory, etc.