Function fvm_sdk::sys::sself::self_destruct

pub unsafe fn self_destruct(burn_funds: bool) -> Result<(), ErrorNumber>

Destroys the calling actor. If burn_funds is true, any unspent balance will be burnt (destroyed). Otherwise, if burnt_funds is false and there are unspent funds, this syscall will fail.

Arguments

burn_funds must be true to delete an actor with unspent funds.

Errors

Error	Reason

IllegalOperation	the actor has unspent funds

ReadOnly	the actor is executing in read-only mode