Module fvm_sdk::vm

Constants

NO_DATA_BLOCK_ID

BlockID representing nil parameters or return data.

Functions

abort

Abort execution; exit code must be non zero.

exit

Exit from current message execution, with the specified code and an optional message and data.

read_only

Returns true if the invocation context is read-only. In read-only mode:

set_panic_handler

Sets a panic handler to turn all panics into aborts with USR_ASSERTION_FAILED. This should be called early in the actor to improve debuggability.