Module fvm_sdk::sys::ipld

Syscalls for manipulating IPLD state.

Structs

IpldOpen

IpldStat

Constants

UNIT

The ID of the “unit” block (or void for C programmers).

Functions

block_create⚠

Creates a new block, returning the block’s ID. The block’s children must be in the reachable set. The new block isn’t added to the reachable set until the CID is computed.

block_link⚠

Computes the given block’s CID, writing the resulting CID into cid.

block_open⚠

Opens a block from the “reachable” set, returning an ID for the block, its codec, and its size in bytes.

block_read⚠

Reads the block identified by id into obuf, starting at offset, reading at most max_len bytes.

block_stat⚠

Returns the codec and size of the specified block.