Function fvm_sdk::sys::ipld::block_read

pub unsafe fn block_read(
    id: u32,
    offset: u32,
    obuf: *mut u8,
    max_len: u32
) -> Result<i32, ErrorNumber>

Reads the block identified by id into obuf, starting at offset, reading at most max_len bytes.

Returns the difference between the length of the block and offset + max_len. This can be used to find the end of the block relative to the buffer the block is being read into:

A zero return value means that the block was read into the output buffer exactly.

A positive return value means that that many more bytes need to be read.

A negative return value means that the buffer should be truncated by the return value.

Arguments

id is ID of the block to read.

offset is the offset in the block to start reading.

obuf is the output buffer (in wasm memory) where the FVM will write the block data.

max_len is the maximum amount of block data to read.

Passing a length/offset that exceed the length of the block will not result in an error, but will result in no data being read and a negative return value indicating where the block actually ended (relative to offset + max_len).

Errors

Error	Reason

InvalidHandle	if the handle isn’t known.

IllegalArgument	if the passed buffer isn’t valid, in memory, etc.